+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - Fix .gif preview rendering on message list 
 - Fix thread shown issue after sending message to a channel 
 
+- Remove border related attributes from MessageInputView. Add close button background attribute to MessageInputView.
+
 # Sep 4th, 2020 - 4.2.11-beta-11
 
 - Fix uploading files and capturing images on Android >= 10