+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 - Implement Giphy actions handler
 - Fix .gif preview rendering on message list
 - Remove border related attributes from MessageInputView. Add close button background attribute to MessageInputView.
-- Forbid setting new user when previous one wasn't disconnected. (Resolves channels list updates)
+- Improve setting user in sample app
 
 # Sep 4th, 2020 - 4.2.11-beta-11